refactor(ui): Extract OAuth LogoGroup components
